# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Học cùng nhau API
  version: 3.0.0
  description: |
    Hợp đồng API lõi của LMS Học cùng nhau. Mọi endpoint dưới /api/v1 yêu cầu cookie phiên `hcn_sid`.
    Request không an toàn (POST/PUT/PATCH/DELETE) yêu cầu header `X-CSRF-Token` và `Origin` hợp lệ.
    Lệnh có hậu quả yêu cầu `Idempotency-Key`. Lỗi theo schema `ErrorResponse`.
servers:
  - url: https://hoc.example.edu.vn
...
security:
  - session: []
paths:
...
  /api/v1/admin/offerings:
    post:
      tags: [admin]
      operationId: createOffering
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId, academicYearId, term, code, title]
              properties:
                courseId: { $ref: '#/components/schemas/Uuid' }
                academicYearId: { $ref: '#/components/schemas/Uuid' }
                term: { type: integer, enum: [1, 2] }
                code: { type: string, pattern: '^[A-Z0-9_-]{3,40}$' }
                title: { type: string, minLength: 1, maxLength: 200 }
                classIds: { type: array, items: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '201': { description: Đã tạo, content: { application/json: { schema: { $ref: '#/components/schemas/Offering' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/admin/offerings/{offeringId}/teachers:
    post:
      tags: [admin]
      operationId: assignTeacher
      parameters: [ { $ref: '#/components/parameters/OfferingId' }, { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teacherId]
              properties:
                teacherId: { $ref: '#/components/schemas/Uuid' }
                capabilities: { type: array, items: { type: string, enum: [teach, author, release, review, view] } }
      responses:
        '201': { description: Đã phân công }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/admin/offerings/{offeringId}/enrollments:
    post:
      tags: [admin]
      operationId: enrollLearners
      parameters: [ { $ref: '#/components/parameters/OfferingId' }, { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [learnerIds]
              properties:
                learnerIds: { type: array, minItems: 1, maxItems: 500, items: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200':
          description: Kết quả từng HS
          content:
            application/json:
              schema:
                type: object
                properties:
                  enrolled: { type: array, items: { $ref: '#/components/schemas/Uuid' } }
                  alreadyEnrolled: { type: array, items: { $ref: '#/components/schemas/Uuid' } }
                  rejected: { type: array, items: { type: object, properties: { learnerId: { $ref: '#/components/schemas/Uuid' }, reason: { type: string } } } }
  /api/v1/admin/users/import:
    post:
      tags: [admin]
      operationId: importUsers
      description: Nhận CSV, tạo tài khoản Keycloak và users. Trả mật khẩu tạm một lần.
      parameters: [ { $ref: '#/components/parameters/Csrf' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } }
      responses:
        '200':
          description: Kết quả
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: array, items: { type: object, properties: { userId: { $ref: '#/components/schemas/Uuid' }, username: { type: string }, temporaryPassword: { type: string } } } }
                  errors: { type: array, items: { type: object, properties: { row: { type: integer }, code: { type: string }, message: { type: string } } } }
                  passwordsRedacted: { type: boolean }
  /api/v1/admin/guardian-links:
    get:
      tags: [admin]
      operationId: listGuardianLinks
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, verified, revoked] } }
      responses:
        '200': { description: Danh sách, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/GuardianLink' } } } } }
    post:
      tags: [admin]
      operationId: createGuardianLink
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [guardianId, learnerId, relation]
              properties:
                guardianId: { $ref: '#/components/schemas/Uuid' }
                learnerId: { $ref: '#/components/schemas/Uuid' }
                relation: { type: string, enum: [father, mother, guardian, other] }
      responses:
        '201': { description: Tạo ở trạng thái pending, content: { application/json: { schema: { $ref: '#/components/schemas/GuardianLink' } } } }
  /api/v1/admin/guardian-links/{linkId}/verify:
    post:
      tags: [admin]
      operationId: verifyGuardianLink
      parameters: [ { $ref: '#/components/parameters/LinkId' }, { $ref: '#/components/parameters/Csrf' } ]
      responses:
        '200': { description: Đã xác minh, content: { application/json: { schema: { $ref: '#/components/schemas/GuardianLink' } } } }
  /api/v1/admin/guardian-links/{linkId}/revoke:
    post:
      tags: [admin]
      operationId: revokeGuardianLink
      parameters: [ { $ref: '#/components/parameters/LinkId' }, { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 3 } } } } }
      responses:
        '200': { description: Đã thu hồi; có hiệu lực ngay request sau }
...
  /api/v1/admin/academic-years:
    get:
      tags: [admin]
      operationId: listAcademicYears
      responses:
        '200': { description: Năm học của trường }
    post:
      tags: [admin]
      operationId: createAcademicYear
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [code, startsOn, endsOn], properties: { code: { type: string }, startsOn: { type: string }, endsOn: { type: string } } } } }
      responses:
        '201': { description: Đã tạo }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/classes:
    get:
      tags: [admin]
      operationId: listAdminClasses
      responses:
        '200': { description: Lớp hành chính }
    post:
      tags: [admin]
      operationId: createAdminClass
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [academicYearId, grade, code], properties: { academicYearId: { $ref: '#/components/schemas/Uuid' }, grade: { type: integer }, code: { type: string } } } } }
      responses:
        '201': { description: Đã tạo }
  /api/v1/admin/classes/{classId}/learners:
    get:
      tags: [admin]
      operationId: listClassLearners
      parameters:
        - { name: classId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200': { description: Học sinh đang học lớp }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/classes/transfer:
    post:
      tags: [admin]
      operationId: transferClass
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [learnerId, toClassId, onDate], properties: { learnerId: { $ref: '#/components/schemas/Uuid' }, toClassId: { $ref: '#/components/schemas/Uuid' }, onDate: { type: string } } } } }
      responses:
        '200': { description: Đã chuyển lớp }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/admin/courses:
    get:
      tags: [admin]
      operationId: listCourses
      responses:
        '200': { description: Môn học của trường }
    post:
      tags: [admin]
      operationId: createCourse
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [subjectCode, grade, title], properties: { subjectCode: { type: string }, grade: { type: integer }, title: { type: string } } } } }
      responses:
        '201': { description: Đã tạo }
  /api/v1/admin/offerings/{offeringId}:
    get:
      tags: [admin]
      operationId: getAdminOffering
      parameters: [ { $ref: '#/components/parameters/OfferingId' } ]
      responses:
        '200': { description: Chi tiết phân công và ghi danh }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/offerings/{offeringId}/assignments/{assignmentId}/end:
    post:
      tags: [admin]
      operationId: endTeacherAssignment
      parameters:
        - { $ref: '#/components/parameters/OfferingId' }
        - { name: assignmentId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - { $ref: '#/components/parameters/Csrf' }
      responses:
        '200': { description: Đã kết thúc phân công }
  /api/v1/admin/offerings/{offeringId}/enrollments/withdraw:
    post:
      tags: [admin]
      operationId: withdrawEnrollment
      parameters: [ { $ref: '#/components/parameters/OfferingId' }, { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [learnerId], properties: { learnerId: { $ref: '#/components/schemas/Uuid' } } } } }
      responses:
        '200': { description: Đã thôi ghi danh }
  /api/v1/admin/users:
    get:
      tags: [admin]
      operationId: listSchoolUsers
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
        - { name: classId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200': { description: Danh sách người dùng, phân trang con trỏ }
  /api/v1/admin/users/{userId}/lock:
    post:
      tags: [admin]
      operationId: lockUser
      parameters:
        - { name: userId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 3 } } } } }
      responses:
        '200': { description: Đã khóa }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/users/{userId}/unlock:
    post:
      tags: [admin]
      operationId: unlockUser
      parameters:
        - { name: userId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 3 } } } } }
      responses:
        '200': { description: Đã mở khóa }
  /api/v1/admin/users/{userId}/reset-password:
    post:
      tags: [admin]
      operationId: resetTemporaryPassword
      parameters:
        - { name: userId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
        - { $ref: '#/components/parameters/Csrf' }
      responses:
        '200':
          description: Mật khẩu tạm, không lưu, Cache-Control no-store
          headers:
            Cache-Control: { schema: { type: string } }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: hcn_sid
  parameters:
    Csrf: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 128 } }
...
    OfferingId: { name: offeringId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
    LinkId: { name: linkId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
  responses:
    Error:
      description: Lỗi
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  schemas:
    Uuid: { type: string, format: uuid }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum: [BAD_REQUEST, UNAUTHENTICATED, FORBIDDEN, CSRF_FAILED, NOT_FOUND, REVISION_CONFLICT, IDEMPOTENCY_KEY_REUSED, REQUEST_IN_PROGRESS,
                     SUBMISSION_VERSION_CHANGED, ALREADY_PUBLISHED, ATTEMPT_LIMIT_REACHED, RELEASE_CLOSED, FILE_TOO_LARGE, FILE_TYPE_NOT_ALLOWED,
                     VALIDATION_FAILED, COVERAGE_BLOCKED, FEATURE_NOT_ENABLED, KC_EDGE_CYCLE, FILE_NOT_SCANNED, RATE_LIMITED, INTERNAL, ASSIGNMENT_OVERLAP]
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    Offering:
      type: object
      required: [id, code, title, subjectCode, grade, term, myRole]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        code: { type: string }
        title: { type: string }
        subjectCode: { type: string }
        grade: { type: integer }
        term: { type: integer }
        myRole: { type: string, enum: [teacher, student, admin] }
    GuardianLink:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        guardianId: { $ref: '#/components/schemas/Uuid' }
        guardianName: { type: string }
        learnerId: { $ref: '#/components/schemas/Uuid' }
        learnerName: { type: string }
        relation: { type: string }
        status: { type: string, enum: [pending, verified, revoked] }
...