# packages/domain/src/index.ts
export { normalizeNumber } from './quiz/normalize.ts'
export { gradeResponse } from './quiz/grade.ts'
export { WEIGHTS, observationFromCriterion, observationFromResponse } from './insight/observe.ts'
export { R0_VERSION, r0Estimate } from './insight/r0.ts'
export { rootGaps } from './insight/rootGaps.ts'
export { computeCoverage } from './authoring/coverage.ts'
export { DomainError, type ErrorCode } from './errors.ts'
export { safeReturnTo } from './identity/returnTo.ts'
export { can, authorize, type Actor, type Action, type Facts, type Decision, type Role } from './identity/policies.ts'
export { getMe } from './identity/getMe.ts'
export { switchContext } from './identity/switchContext.ts'
export { establishSession, revokeSession } from './identity/establishSession.ts'
export { systemClock, type Clock } from './org/clock.ts'
export { IdpAdminError, type IdpAdmin, type CreateIdpUser, type IdpUser } from './org/idp-admin.ts'
export { temporaryPassword, temporaryPasswordAlphabet } from './org/temporary-password.ts'
export { createAcademicYear, listAcademicYears } from './org/createAcademicYear.ts'
export { createAdminClass, listAdminClasses, listClassLearners } from './org/createAdminClass.ts'
export { createCourse, listCourses } from './org/createCourse.ts'
export { createOffering, type OfferingDto } from './org/createOffering.ts'
export { assignTeacher } from './org/assignTeacher.ts'
export { endTeacherAssignment } from './org/endTeacherAssignment.ts'
export { enrollLearners } from './org/enrollLearners.ts'
export { withdrawEnrollment } from './org/withdrawEnrollment.ts'
export { createGuardianLink, listGuardianLinks, type GuardianLinkDto } from './org/createGuardianLink.ts'
export { verifyGuardianLink } from './org/verifyGuardianLink.ts'
export { revokeGuardianLink } from './org/revokeGuardianLink.ts'
export { transferClass } from './org/transferClass.ts'
export { listSchoolUsers, type SchoolUser } from './org/listSchoolUsers.ts'
export { lockUser } from './org/lockUser.ts'
export { unlockUser } from './org/unlockUser.ts'
export { resetTemporaryPassword } from './org/resetTemporaryPassword.ts'
export { importUsers, type ImportResult } from './org/importUsers.ts'
export { listMyOfferings, getOffering, getAdminOffering } from './org/listMyOfferings.ts'
export { listMyChildren, getChildOverview } from './org/listMyChildren.ts'
export { seedCurriculum, type SeedCurriculumResult } from './org/seedCurriculum.ts'
export { bootstrapSchool } from './org/bootstrapSchool.ts'